+++ mail_2_rmq.py
@@ -29,35 +29,38 @@
         RabbitMQ configuration file format (config/rabbitmq.py.TEMPLATE).
 
             # RabbitMQ Configuration file
-            # Classification (U)
-            # Unclassified until filled.
             user = "USER"
-            passwd = "PASSWORD"
+            japd = "PASSWORD"
             host = "HOSTNAME"
-            # RabbitMQ listening port, default is 5672.
-            port = 5672
             # RabbitMQ Exchange name for each instance run.
             exchange_name = "EXCHANGE_NAME"
+            # List of valid queues in RabbitMQ.
+            # Note:  Queues names must be UpperCamelCase style.
+            valid_queues = ["QueueName1", "QueueName2", ...]
+            # Name of error queue to handle incorrectly routed emails.
+            err_queue = "ERROR_QUEUE_NAME"
+            # Archive directory path for non-processed email files.
+            email_dir = "DIRECTORY_PATH/email_dir"
+            # Directory path and file name to the program log.
+            log_file = "DIRECTORY_PATH/logs/mail_2_rmq.log"
+            # Directory path to temporary staging directory.
+            tmp_dir = "DIRECTORY_PATH/tmp"
+            # Do not modify settings below unless you know what you are doing.
+            # Filter out strings within the subject line.
+            # Do not modify unless you understand regular expressions.
+            subj_filter = ["\[.*\]"]
+            # Types of attachments to extract from email.
+            attach_types = ["application/pdf"]
+            # RabbitMQ listening port, default is 5672.
+            port = 5672
             # Type of exchange:  direct, topic, fanout, headers
             exchange_type = "direct"
             # Is exchange durable: True|False
             x_durable = True
             # Are queues durable: True|False
             q_durable = True
-            # Do queues delete once message is processed:  True|False
+            # Do queues automatically delete once message is processed.
             auto_delete = False
-            # List of valid queues in RabbitMQ.
-            valid_queues = ["QUEUE_NAME1", "QUEUE_NAME2", ... ]
-            # Name of error queue to handle incorrect email subjects.
-            err_queue = "ERROR_QUEUE_NAME"
-            # Archive directory path for non-processed email files.
-            email_dir = "DIRECTORY_PATH/email_dir"
-            # Directory path and file name to the program log.
-            log_file = "DIRECTORY_PATH/logs/mail_2_rmq.log"
-            # Filter out strings within the subject line.
-            # Do not modify this setting unless you understand regular
-            #   expressions.
-            subj_filter = ["\[.*\]"]
 
     Example:
         alias: "| /opt/local/mail_2_rmq.py -M -c rabbitmq -d /opt/local/config"