+++ test/blackbox/mail_2_rmq/mail_2_rmq_cleanup.py
@@ -109,7 +109,7 @@ def mail_2_rmq_cleanup(cfg, queue_name, drop_exch=False):
     """
 
     rmq = rabbitmq_class.RabbitMQPub(
-        cfg.user, cfg.passwd, cfg.host, cfg.port, cfg.exchange_name,
+        cfg.user, cfg.japd, cfg.host, cfg.port, cfg.exchange_name,
         cfg.exchange_type, queue_name, queue_name, cfg.x_durable,
         cfg.q_durable, cfg.auto_delete)