+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 - config/rabbitmq.py.TEMPLATE:  Added file_queue, tmp_dir, and attach_types entries.
 
 ### Changed
+- run_program:  Setup flavor_id to use -y option value or exchange_name as a default.
 - main:  Added -y option to the setup.
 - run_program:  Refactored status_flag check.
 - create_rq:  Changed configuration settings to reflect changes in config file.